more

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/MecanumDrive.java
@@ -158,10 +158,10 @@ public class MecanumDrive {
             if (!initialized) {
                 initialized = true;
 
-                lastLeftFrontPos = leftFrontPosVel.position;
-                lastLeftBackPos = leftBackPosVel.position;
-                lastRightBackPos = rightBackPosVel.position;
-                lastRightFrontPos = rightFrontPosVel.position;
+                lastLeftFrontPos = (int) leftFrontPosVel.position;
+                lastLeftBackPos = (int) leftBackPosVel.position;
+                lastRightBackPos = (int) rightBackPosVel.position;
+                lastRightFrontPos = (int) rightFrontPosVel.position;
 
                 lastHeading = heading;
 
@@ -191,10 +191,10 @@ public class MecanumDrive {
                     }).times(PARAMS.inPerTick)
             ));
 
-            lastLeftFrontPos = leftFrontPosVel.position;
-            lastLeftBackPos = leftBackPosVel.position;
-            lastRightBackPos = rightBackPosVel.position;
-            lastRightFrontPos = rightFrontPosVel.position;
+            lastLeftFrontPos = (int) leftFrontPosVel.position;
+            lastLeftBackPos = (int) leftBackPosVel.position;
+            lastRightBackPos = (int) rightBackPosVel.position;
+            lastRightFrontPos = (int) rightFrontPosVel.position;
 
             lastHeading = heading;
